Added factory class %26 Added example of JPL
I added a static factory class for the zif_timer interface and it uses the zcl_timer class. 
The factory can create a root timer and normal timers, the differences is that root timers don't need a reference to a timer up the hierarchy.
The factory also collects all timers in the mt_jpl_timer_table so all timers can be summarized later

I have also added a new package which includes a test report which will show how to use this library. But this not the focus of the commit and is not done.